Reverted changes
Removed all code from previous two commits.  Noah if you add code to the project don't break it.  We can't work on the game if you put in code that doesn't let us compile.  Also don't write code with dependencies that are not part of this application.  Write what you need, don't just copy what the code does in that tutorial you found.  I'm trying not to lose my mind and you aren't making it easy.